backend/agents/deploy_agent.py
import os
from dotenv import load_dotenv

load_dotenv()
from uagents_core.utils.registration import (
    register_chat_agent,
    RegistrationRequestCredentials,
)

register_chat_agent(
    "ER triage agent",
    "https://agentverse.ai/v1/submit",
    active=True,
    credentials=RegistrationRequestCredentials(
        agentverse_api_key=os.environ["AGENTVERSE_KEY"],
        agent_seed_phrase=os.environ["AGENT_SEED_PHRASE"],
    ),
)

print("Agent registered successfully")