backend/requirements.txt
uagents>=0.14.0
anthropic>=0.39.0
python-dotenv>=1.0.0
chromadb>=0.1.0
flask>=3.0.0
flask-cors>=4.0.0